feat: creating review and confirm screen
