+++ lib/core/routes.dart
@@ -1,4 +1,4 @@
-import 'package:crypto_listing/convert/widgets/check_out_page.dart';
+import 'package:crypto_listing/review/view/review_page.dart';
 import 'package:crypto_listing/convert/widgets/complete_confirmation.dart';
 
 import '../convert/view/convert_screen.dart';
@@ -11,6 +11,6 @@ var routesNavigation = {
   '/criptoMoviments': (context) => const MovementsScreen(),
   '/criptoDetails': (context) => const DetailScreen(),
   '/convert': (context) => const ConvertScreen(),
-  '/confirm-conversion': (context) => const CheckOutPage(),
+  '/confirm-conversion': (context) => const ReviewPage(),
   '/complete-conversion': (context) => const CompleteConfirmation()
 };
