+++ lib/review/view/review_page.dart
@@ -1,0 +1,20 @@
+import 'package:crypto_listing/shared/widgets/default_appbar.dart';
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+import '../widgets/review_body.dart';
+
+class ReviewPage extends ConsumerWidget {
+  const ReviewPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return const Scaffold(
+      appBar: DefaultAppbar(
+        title: 'Revisar',
+        titleSize: 22,
+      ),
+      body: ReviewBody(),
+    );
+  }
+}
